Improve error messages for programmer mistakes.
1. For when an Array is expected but you pass a Hash

Before, its error message came from implicit conversion of the hash key
into an Integer by its use as an index on the Array.

This is to work with the type of XML that is fairly common, e.g.:
 <stallions>
   <stallion>a</stallion>
   <stallion>b</stallion>
   <stallion>c</stallion>
 </stallion>

But it is easy to mess up the format in the render :soap.

2. For when an Array is expected but you pass a String

This was silently discarding the value. Now it raises an exception.