Add Guard for testing.
Now you can run:

    bundle exec guard

And changes to the spec/* or to the lib/* will trigger re-execution of
the tests.

Also bumped the development dep on `pry` to `pry-de`, which includes a
bunch of plugins to do debugging (stepping), stack exploration
(show-stack, up), and so on. You might not like this, but I definitely
do.  Feel free to take that part of the commit out if you don't want it
but do want the Guard part.