Adding support for WSSE Authentication callbacks
Added a `soap_config` parameter that specifies a callback used to
authenticate the user and password of a WSSE secured request.

This lambda approach was preferred so as not to introduce too many
changes in the way WSSE already works, while providing the flexibility
to use a default user/password or a custom mechanism for
authentication.