Set xsi:nil to true if content is nil
Prior to this commit the XML returned by `wash_out` for `nil` values has a empty
element but the type is marked as the type of the expected value. When
clients like Savon parse this XML they return the type instead of the value
which is actually nil.

With this commit if the value is passed in is nil then `xsi:nil` will be set to
`true` on the element instead of the type. Now when clients like Savon parse the
XML they will return the actual given value `nil` instead of the type the
element was supposed to have.

According to the XML standard: xsi:nil is "a mechanism
for signaling that an element should be accepted as ·valid· when it has no
content despite a content type which does not require or even necessarily allow
empty content. An element may be ·valid· without content if it has the attribute
xsi:nil with the value true. An element so labeled must be empty, but can carry
attributes if permitted by the corresponding complex type."

https://www.w3.org/TR/xmlschema-1/#xsi_nil