Handle arrays of empty elements properly
It used to be that arrays containing empty string members returned
their members as stringified hashes of the attributes instead of the
empty string:

  <item>
    <foo xsi:type="xsd:string"></foo>
  </item>
  <item>
    <foo xsi:type="xsd:string"></foo>
  </item>

  expect(list.first["foo"]).to eq("")
  # Is actually: '{"@xsi:type"=>"xsd:string"}`

This is a pretty conservative change as it only kicks in when a
parameter that shouldn't be a struct is a Hash *and* all the keys in
that hash are attributes.

That means that if the element was actually nested elements, it will
still be returned as a stringified Hash.